Move objects and schemas to domain
